# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/carts": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Consultar carritos",
                "description": "Retorna todos los carritos",
                "operationId": "082ccb576b0f3887aeea5ab4ad1e76b8",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Crear carrito",
                "description": "Crea un nuevo carrito",
                "operationId": "3dfdfabc59b4f5159eacaff7b56d91b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carrito creado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/carts/{cart}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Consultar un carrito",
                "description": "Retorna un carrito específico",
                "operationId": "9de67301b4c4fa527da646cdeea2a0f7",
                "parameters": [
                    {
                        "name": "cart",
                        "in": "path",
                        "description": "ID del carrito",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Carrito no encontrado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "Actualizar carrito",
                "description": "Actualiza un carrito existente",
                "operationId": "70f8d79e936832aa230907518b7b2cc5",
                "parameters": [
                    {
                        "name": "cart",
                        "in": "path",
                        "description": "ID del carrito",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Carrito actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Carrito no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Eliminar carrito",
                "description": "Elimina un carrito existente",
                "operationId": "8d7f0d90ac93117163d7e8844fd4c860",
                "parameters": [
                    {
                        "name": "cart",
                        "in": "path",
                        "description": "ID del carrito",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carrito eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Carrito no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}